+++ shell/config/private-label.js
@@ -23,6 +23,21 @@ export function setMode(m) {
 
 export function setVendor(v) {
   vendor = v;
+  document.title = v;
+  if (v === 'Harvester') {
+    const ico = require(`~shell/assets/images/pl/harvester.png`);
+
+    document.title = 'Harvester';
+    const link = document.createElement('link');
+
+    link.hid = 'icon';
+    link.rel = 'icon';
+    link.type = 'image/x-icon';
+    link.hrefv = ico;
+    const head = document.getElementsByTagName('head')[0];
+
+    head.appendChild(link);
+  }
 }
 
 export function setProduct(p) {
