+++ shell/mixins/brand.js
@@ -1,6 +1,5 @@
 import { mapGetters } from 'vuex';
 import { CATALOG, MANAGEMENT } from '@shell/config/types';
-import { getVendor } from '@shell/config/private-label';
 import { SETTING } from '@shell/config/settings';
 import { findBy } from '@shell/utils/array';
 import { createCssVars } from '@shell/utils/color';
@@ -116,6 +115,7 @@ export default {
       if (this.linkColor) {
         this.setCustomColor(this.linkColor, 'link');
       }
+      this.setBodyClass();
     },
 
     cspAdapter(neu) {
@@ -142,45 +142,14 @@ export default {
           }
         }
       }
+    },
+    brand() {
+      this.setBodyClass();
     }
+
   },
   mounted() {
-    const cssClass = `overflow-hidden dashboard-body`;
-    const body = document.getElementsByTagName('body')[0];
-    let title = getVendor();
-    let bodyClass, brandMeta;
-
-    if (getVendor() === 'Harvester') {
-      const ico = require(`~shell/assets/images/pl/harvester.png`);
-
-      title = 'Harvester';
-      const link = document.createElement('link');
-
-      link.hid = 'icon';
-      link.rel = 'icon';
-      link.type = 'image/x-icon';
-      link.hrefv = ico;
-      const head = document.getElementsByTagName('head')[0];
-
-      head.appendChild(link);
-    }
-
-    if ( this.brand ) {
-      try {
-        brandMeta = require(`~shell/assets/brand/${ this.brand }/metadata.json`);
-        if (brandMeta?.hasStylesheet === 'true') {
-          bodyClass = `${ cssClass } ${ this.brand } theme-${ this.theme }`;
-        } else {
-          bodyClass = `theme-${ this.theme } overflow-hidden dashboard-body`;
-          this.$store.dispatch('prefs/setBrandStyle', this.theme === 'dark');
-        }
-      } catch {
-        bodyClass = `theme-${ this.theme } ${ cssClass }`;
-      }
-    }
-
-    document.title = title;
-    body.className = bodyClass;
+    this.setBodyClass();
   },
   methods: {
     setCustomColor(color, name = 'primary') {
@@ -197,6 +166,25 @@ export default {
       for (const prop in vars) {
         document.body.style.removeProperty(prop);
       }
+    },
+    setBodyClass() {
+      const body = document.getElementsByTagName('body')[0];
+      const cssClass = `overflow-hidden dashboard-body`;
+      let bodyClass = `theme-${ this.theme } ${ cssClass }`;
+
+      if ( this.brand ) {
+        try {
+          const brandMeta = require(`~shell/assets/brand/${ this.brand }/metadata.json`);
+
+          if (brandMeta?.hasStylesheet === 'true') {
+            bodyClass = `${ cssClass } ${ this.brand } theme-${ this.theme }`;
+          } else {
+            bodyClass = `theme-${ this.theme } overflow-hidden dashboard-body`;
+            this.$store.dispatch('prefs/setBrandStyle', this.theme === 'dark');
+          }
+        } catch {}
+      }
+      body.className = bodyClass;
     }
   }
